+++ train_keras_model.py
@@ -22,7 +22,7 @@ import training_data
 def choose_action(model, observation, epsilon=0.1):
     """Choose best action from the esimator or random, based on epsilon
        Return both the action id and the estimated quality."""
-    predictions = np.reshape(model.predict(np.reshape(observation, (-1, 256))), (4, ))
+    predictions = np.reshape(model.predict(np.reshape(observation.astype('float32'), (-1, 256))), (4, ))
     #print(predictions)
     if random.uniform(0, 1) > epsilon:
         chosen = np.argmax(predictions)
